# generator: Replace `Extends{Root}` with `Extends<Root>` trait