--- a/ash/src/lib.rs
+++ b/ash/src/lib.rs
@@ -268,3 +268,9 @@ where
         }
     }
 }
+
+#[test]
+fn trybuild() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("tests/fail/long_lived_root_struct_borrow.rs");
+}