--- a/ash/tests/fail/long_lived_root_struct_borrow.stderr
+++ b/ash/tests/fail/long_lived_root_struct_borrow.stderr
@@ -1,0 +1,23 @@
+error[E0502]: cannot borrow `api` as immutable because it is also borrowed as mutable
+  --> tests/fail/long_lived_root_struct_borrow.rs:10:10
+   |
+7  |     let pdev_props = vk::PhysicalDeviceProperties2::default().push(&mut api);
+   |                                                                    -------- mutable borrow occurs here
+...
+10 |     dbg!(&api);
+   |          ^^^^ immutable borrow occurs here
+...
+13 |     dbg!(pdev_props); // Holds a borrow on api and layers
+   |          ---------- mutable borrow later used here
+
+error[E0502]: cannot borrow `layers` as immutable because it is also borrowed as mutable
+  --> tests/fail/long_lived_root_struct_borrow.rs:11:10
+   |
+6  |         vk::PhysicalDeviceLayeredApiPropertiesListKHR::default().layered_apis(&mut layers);
+   |                                                                               ----------- mutable borrow occurs here
+...
+11 |     dbg!(&layers);
+   |          ^^^^^^^ immutable borrow occurs here
+12 |
+13 |     dbg!(pdev_props); // Holds a borrow on api and layers
+   |          ---------- mutable borrow later used here