@MarijnS95 2024-12-08
It looks like this PR comes with a ±150ms build speedup:

## `master` (92a0ee6)

```console
$ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.964 s ±  0.109 s    [User: 7.768 s, System: 0.902 s]
  Range (min … max):    7.797 s …  8.133 s    10 runs
```

## This PR (1cf5047d)

```console
$ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.817 s ±  0.092 s    [User: 7.669 s, System: 0.845 s]
  Range (min … max):    7.707 s …  8.008 s    10 runs
```