@Ralith 2024-12-08
Awesome! Glad we finally stumbled on *something* that helps there, and great that it's a readability/concision win too.