@MarijnS95 2024-12-08
Sneak-preview: moving the functions to `AnyTaggedStructure` is going to give us another ±90ms :)