@MarijnS95 2024-12-09
Performance optimization is now less significant, ±105ms, perhaps because of the expanded lifetimes?

(Retested `master` too and its performance measurement was within margin of error on my machine)

## This PR (4305ec7d):

```console
❯ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.858 s ±  0.103 s    [User: 7.702 s, System: 0.876 s]
  Range (min … max):    7.745 s …  8.026 s    10 runs
```