@MarijnS95 2025-05-08
### `master` @ 98e5db8

```console
❯ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.118 s ±  0.072 s    [User: 7.144 s, System: 0.827 s]
  Range (min … max):    7.041 s …  7.233 s    10 runs
```

### This PR @ cbdf201

```console
❯ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.149 s ±  0.093 s    [User: 7.168 s, System: 0.854 s]
  Range (min … max):    7.054 s …  7.352 s    10 runs
```

### This PR @ cbdf201 with unnamed (`'_`) lifetime

```rust
quote!(unsafe impl Extends<#base<'_>> for #name<'_> {})
```
Yields:
```console
❯ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.141 s ±  0.086 s    [User: 7.163 s, System: 0.832 s]
  Range (min … max):    7.011 s …  7.271 s    10 runs
```

### `'a`

Just in case I named the parameter with:

```rust
quote!(unsafe impl<'a> Extends<#base<'a>> for #name<'a> {})
```
Within margin of error:
```console
❯ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.155 s ±  0.086 s    [User: 7.187 s, System: 0.834 s]
  Range (min … max):    7.046 s …  7.352 s    10 runs
```

### Explicitly specifying the `'child: 'root` bound:

```rust
quote!(unsafe impl<'root, 'child: 'root> Extends<#base<'root>> for #name<'child> {})
```

```console
❯ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      7.143 s ±  0.090 s    [User: 7.157 s, System: 0.847 s]
  Range (min … max):    6.977 s …  7.261 s    10 runs
```