@MarijnS95 2025-05-08
Just in case I prefer the nameless `'_` variant, because it keeps a lot more `impl`s on the same line. We have tests ensuring that the lifetime bounds work as we expect (allowing certain cases, while failing others).

```console
❯ git diff --stat
 ash/src/vk/definitions.rs | 3991 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++-------------------------------------------------------------------------------------------------------------------------------------------------
 generator/src/lib.rs      |    2 +-
 2 files changed, 1215 insertions(+), 2778 deletions(-)
```