@Ralith 2025-05-08
So in summary, no obvious impact from anything? Weird, I wonder what we measured before. Regardless, I'm happy to proceed with the terse anonymous lifetime variant.